up-to-date

+++ CNProject/src/Message.java
@@ -2,10 +2,10 @@ import java.io.Serializable;
 
 public class Message implements Serializable {		
 	int length ;
-	int type;
+	byte type;
 	byte[] payload;
 	
-	public Message(int type , int payloadsize)
+	public Message(byte type , int payloadsize)
 	{
 		this.type = type;
 		this.payload = new byte[payloadsize];
